==> Extention_all_IdentityTerms/7_analysis_of_results.ipynb <==
This notebook contains the the results for necessity and sufficiency. Necessity and sufficiency are both calculated by either choosing a subset of tokens and perturbing them using the ILM model. The models are all BERT architecture, but trained on different datasets, and for each dataset, a model is trained on both hate/non-hate and abusive/non-abusive labels. The explanations are generated for 120 examples from the HateCheck test suite. These are instances that are explicitly hateful, and are targeted towards women or Muslims. The function ```display_scores``` displays the necessity and sufficiency for each of the examples for all models included. Note that some models will display ```NaN``` for some values. These are the cases where the model mistakenly classified the original instance as non-abusive/non-hateful. In these cases, the current necessity and sufficiency calculations aren't meaningful, because we aim to provide explanations for positive predictions only. The third argument to this function determines which necessity/sufficiency scores to display. 

import pickle
import pandas as pd
import numpy as np

preds = pickle.load(open("Data/HateCheck_necc_suff_preds_2.pickle", "rb"))
results = pickle.load(open("Data/HateCheck_necc_suff_results_all_2.pickle", "rb"))
perturbations = pickle.load(open("Data/HateCheck_necc_suff_perturbations_3.pickle","rb"))

perturbations.keys()

preds.keys()

results.keys()

datasets = list(results['necc_results'].keys())
datasets

# get the corrupted examples with tokn k flipped together with the corresponding scores
# if this is sufficiency, set reverse=True so that it will return instances where token k is not flipped
def get_k_corr(k, masks, perturbed, p_results, reverse=False):
    perturbed_k = []
    for pp, mm, rr in zip(perturbed, masks[:,k], p_results):
        if mm != reverse:
            perturbed_k.append((pp, rr))
    return(perturbed_k)

# for a given original test case, make a table for necessity or sufficiency for each model and for each token

def display_scores(templ_n, orig_texts, orig_preds, scores_dict):
    columns = orig_texts[templ_n].strip().split()
    index = list(scores_dict.keys())
    # if the model has not predicted the original as 0, do not display the feature attribution scores
    data = np.array([scores_dict[dset][templ_n] if orig_preds[dset][templ_n] == 1 
                     else [np.nan]*len(scores_dict[dset][templ_n]) 
                    for dset in scores_dict.keys()])
    return pd.DataFrame(data=data, columns=columns, index=index)

display_scores(0, perturbations['orig_texts'],  preds['orig_preds'], results['necc_results'])

display_scores(0, perturbations['orig_texts'],  preds['orig_preds'], results['suff_results'])

display_scores(1, perturbations['orig_texts'],  preds['orig_preds'], results['necc_results'])

display_scores(1, perturbations['orig_texts'],  preds['orig_preds'], results['suff_results'])

# This function is to examine the perturbations and the scores a model assigned them. 
# Will not work for masked-not-perturbed versions. 

ex_no = 0 # example in the examples list
id_token = 2  #the identity token is the _th token
model = 'CAD_abuse'
get_k_corr(id_token, 
           perturbations['suff_masks'][ex_no], 
           perturbations['suff_perturbed'][ex_no], 
           preds['suff_scores'][model][ex_no], reverse=True)

ex_no = 0 # example in the examples list
id_token = 2  #the identity token is the _th token
model = 'CAD_abuse'
get_k_corr(id_token, 
           perturbations['necc_masks'][ex_no], 
           perturbations['necc_perturbed'][ex_no], 
           preds['necc_scores'][model][ex_no])

In the following section, there is the average prediction of the models for the cases where identities are mentioned in neutral or positive contexts (```ident_neutral_nh```, ```ident_pos_nh```) and the test cases where there is "abuse" directed to non-protected groups (```target_group_nh```) individuals (```target_indiv_nh```) and objects (```target_obj_nh```). The performance of the models on non-abusive mention identities for the two targets we've chosen for our explanations ('women' and 'Muslims') is also given seperately. 

hc_results = pickle.load(open('Data/HateCheck_templates_and_results_2.pickle', "rb"))
hc_results.test_case = hc_results.test_case.apply(lambda x: x.strip())
hc_results

# target_ds = ['women', 'trans people', 'gay people', 'black people', 'disabled people',' Muslims', 'immigrants']

hc_results_women_nh = hc_results[((hc_results.functionality == 'ident_neutral_nh') 
                         | (hc_results.functionality == 'ident_pos_nh')) 
                        & (hc_results.target_ident == 'women')]

hc_results_trans_nh = hc_results[((hc_results.functionality == 'ident_neutral_nh') 
                         | (hc_results.functionality == 'ident_pos_nh')) 
                        & (hc_results.target_ident == 'trans people')]

hc_results_gay_nh = hc_results[((hc_results.functionality == 'ident_neutral_nh') 
                         | (hc_results.functionality == 'ident_pos_nh')) 
                        & (hc_results.target_ident == 'gay people')]

hc_results_black_nh = hc_results[((hc_results.functionality == 'ident_neutral_nh') 
                         | (hc_results.functionality == 'ident_pos_nh')) 
                        & (hc_results.target_ident == 'black people')]

hc_results_disabled_nh = hc_results[((hc_results.functionality == 'ident_neutral_nh') 
                         | (hc_results.functionality == 'ident_pos_nh')) 
                        & (hc_results.target_ident == 'disabled people')]

hc_results_muslims_nh = hc_results[((hc_results.functionality == 'ident_neutral_nh') 
                         | (hc_results.functionality == 'ident_pos_nh')) 
                        & (hc_results.target_ident == 'Muslims')]

hc_results_immigrants_nh = hc_results[((hc_results.functionality == 'ident_neutral_nh') 
                         | (hc_results.functionality == 'ident_pos_nh')) 
                        & (hc_results.target_ident == 'immigrants')]


# hc_results.functionality.loc[((hc_results.functionality == 'ident_neutral_nh') 
#                          | (hc_results.functionality == 'ident_pos_nh')) 
#                         & (hc_results.target_ident == 'women')] = 'women_nh'
# hc_results.functionality.loc[((hc_results.functionality == 'ident_neutral_nh') 
#                          | (hc_results.functionality == 'ident_pos_nh')) 
#                         & (hc_results.target_ident == 'Muslims')] = 'muslims_nh'

mask = (
        (hc_results["functionality"] == "ident_neutral_nh") | 
        (hc_results["functionality"] == "ident_pos_nh")
    ) & (hc_results["target_ident"] == "women")
hc_results.loc[mask, "functionality"] = "women_nh"

mask = (
        (hc_results["functionality"] == "ident_neutral_nh") | 
        (hc_results["functionality"] == "ident_pos_nh")
    ) & (hc_results["target_ident"] == 'trans people')
hc_results.loc[mask, "functionality"] = "trans_nh"

mask = (
        (hc_results["functionality"] == "ident_neutral_nh") | 
        (hc_results["functionality"] == "ident_pos_nh")
    ) & (hc_results["target_ident"] == 'gay people')
hc_results.loc[mask, "functionality"] = "gay_nh"

mask = (
        (hc_results["functionality"] == "ident_neutral_nh") | 
        (hc_results["functionality"] == "ident_pos_nh")
    ) & (hc_results["target_ident"] == 'black people')
hc_results.loc[mask, "functionality"] = "black_nh"

mask = (
        (hc_results["functionality"] == "ident_neutral_nh") | 
        (hc_results["functionality"] == "ident_pos_nh")
    ) & (hc_results["target_ident"] == 'disbaled people')
hc_results.loc[mask, "functionality"] = "disabled_nh"

mask = (
        (hc_results["functionality"] == "ident_neutral_nh") | 
        (hc_results["functionality"] == "ident_pos_nh")
    ) & (hc_results["target_ident"] == 'Muslims')
hc_results.loc[mask, "functionality"] = "muslims_nh"

mask = (
        (hc_results["functionality"] == "ident_neutral_nh") | 
        (hc_results["functionality"] == "ident_pos_nh")
    ) & (hc_results["target_ident"] == 'immigrants')
hc_results.loc[mask, "functionality"] = "immigrants_nh"
# target_ds = ['women', 'trans people', 'gay people', 'black people', 'disabled people',' Muslims', 'immigrants']


# the results we are interested are: 
target_funcs = ['women_nh','trans_nh', 'gay_nh', 'black_nh', 'disabled_nh', 'muslims_nh', 'immigrants_nh', 
                'target_obj_nh', 'target_indiv_nh', 'target_group_nh']

target_funcs_results = hc_results[hc_results.functionality.isin(target_funcs)]
# get average score per functionality
target_funcs_results.groupby('functionality')[['{}_pred'.format(dd) for dd in datasets]].mean().transpose()

mask_results = pickle.load(open('Data/HateCheck_necc_suff_results_masked_2.pickle', 'rb'))
mask_results.keys()

necc_vals = {}
suff_vals = {}
necc_vals_mask = {}
suff_vals_mask = {}
orig_texts = []
targets = []

for tt in perturbations['orig_texts']:
    orig_text = tt.strip()
    row = hc_results[hc_results.test_case == orig_text]
    targets.append(row.target_ident.tolist()[0])

for dataset in datasets:
    necc_vals[dataset] = []
    suff_vals[dataset] = []
    necc_vals_mask[dataset] = []
    suff_vals_mask[dataset] = []
    for nn, (orig_text, orig_pred) in enumerate(zip(perturbations['orig_texts'], preds['orig_preds'][dataset])):
        if orig_pred != 1:
            necc_vals[dataset].append(np.nan)
            suff_vals[dataset].append(np.nan)
            necc_vals_mask[dataset].append(np.nan)
            suff_vals_mask[dataset].append(np.nan)
            continue
        # get the row in hc_results corresponding to this case
        orig_text = orig_text.strip()
        row = hc_results[hc_results.test_case == orig_text]
        toknd = row.case_templ.tolist()[0].split()
        ## find the index of the template placeholder
        for ii, tt in enumerate(toknd):
            if tt[:1] == "[":
                break
        necc_vals[dataset].append(results['necc_results'][dataset][nn][ii])
        suff_vals[dataset].append(results['suff_results'][dataset][nn][ii])
        necc_vals_mask[dataset].append(mask_results['necc_results_nb'][dataset][nn][ii])
        suff_vals_mask[dataset].append(mask_results['suff_results_nb'][dataset][nn][ii])

df_dict = {('necessity', dd): ll for dd, ll in necc_vals.items()}
df_dict.update({('sufficiency', dd): ll for dd, ll in suff_vals.items()})
df_dict.update({('necessity_mask', dd): ll for dd, ll in necc_vals_mask.items()})
df_dict.update({('sufficiency_mask', dd): ll for dd, ll in suff_vals_mask.items()})
df_dict.update({('prediction', dd): ll for dd, ll in preds['orig_preds'].items()})
df_dict.update({('score', dd): ll for dd, ll in preds['orig_scores'].items()})
#df_dict.update({'target', ''}: targets)

#ind = [xx.strip() for xx in perturbations['orig_texts']]
ind = [(tt, xx.strip()) for xx, tt in zip(perturbations['orig_texts'], targets)]

# pd.DataFrame(df_dict, index=ind)
#     avg_necc[dataset] = {target: np.mean(necc_vals[target]) for target in targets}
#     avg_suff[dataset] = {target: np.mean(suff_vals[target]) for target in targets}

master_df = pd.DataFrame(df_dict, index=ind)
master_df.columns = pd.MultiIndex.from_tuples(master_df.columns, names=['value','Dataset'])
master_df.index = pd.MultiIndex.from_tuples(master_df.index, names=['target', 'text'])
pickle.dump(master_df, open("Data/HateCheck_individual_necc_suff_scores_2.pickle", "wb"))

# master_df.xs('CAD_abuse', level='Dataset', axis=1)
# master_df['necessity']
# master_df.loc['women']
# master_df.xs('I hate women.', level='text')

master_df = pickle.load(open("Data/HateCheck_individual_necc_suff_scores_2.pickle", "rb"))

master_df['necessity'].groupby(level='target').mean().transpose()

master_df['necessity'].groupby(level='target').std().transpose()

master_df['sufficiency'].groupby(level='target').mean().transpose()

master_df['sufficiency'].groupby(level='target').std().transpose()

master_df['necessity_mask'].groupby(level='target').mean().transpose()

master_df['sufficiency_mask'].groupby(level='target').mean().transpose()
